# [exporter/splunkhec] apply multimetric metric merge for the whole batch