--- a/exporter/splunkhecexporter/client.go
+++ b/exporter/splunkhecexporter/client.go
@@ -4,6 +4,7 @@
 package splunkhecexporter // import "github.com/open-telemetry/opentelemetry-collector-contrib/exporter/splunkhecexporter"
 
 import (
+	"bytes"
 	"context"
 	"errors"
 	"fmt"
@@ -96,6 +97,9 @@ func (c *client) pushMetricsData(
 		}
 	}
 
+	if c.config.UseMultiMetricFormat {
+		return c.pushMultiMetricsDataInBatches(ctx, md, localHeaders)
+	}
 	return c.pushMetricsDataInBatches(ctx, md, localHeaders)
 }
 
@@ -242,20 +246,56 @@ func (c *client) fillMetricsBuffer(metrics pmetric.Metrics, buf buffer, is iterS
 	jsonStream := jsonStreamPool.Get().(*jsoniter.Stream)
 	defer jsonStreamPool.Put(jsonStream)
 
-	var events []*splunk.Event
-
-	for i := 0; i < metrics.ResourceMetrics().Len(); i++ {
+	for i := is.resource; i < metrics.ResourceMetrics().Len(); i++ {
 		rm := metrics.ResourceMetrics().At(i)
-		for j := 0; j < rm.ScopeMetrics().Len(); j++ {
+		for j := is.library; j < rm.ScopeMetrics().Len(); j++ {
+			is.library = 0 // Reset library index for next resource.
 			sm := rm.ScopeMetrics().At(j)
-			for k := 0; k < sm.Metrics().Len(); k++ {
+			for k := is.record; k < sm.Metrics().Len(); k++ {
+				is.record = 0 // Reset record index for next library.
 				metric := sm.Metrics().At(k)
 
 				// Parsing metric record to Splunk event.
-				events = append(events, mapMetricToSplunkEvent(rm.Resource(), metric, c.config, c.logger)...)
+				events := mapMetricToSplunkEvent(rm.Resource(), metric, c.config, c.logger)
+				tempBuf := bytes.NewBuffer(make([]byte, 0, c.config.MaxContentLengthMetrics))
+				for _, event := range events {
+					// JSON encoding event and writing to buffer.
+					b, err := marshalEvent(event, c.config.MaxEventSize, jsonStream)
+					if err != nil {
+						permanentErrors = append(permanentErrors, consumererror.NewPermanent(fmt.Errorf("dropped metric event: %v, error: %w", event, err)))
+						continue
+					}
+					tempBuf.Write(b)
+				}
+
+				// Continue adding events to buffer up to capacity.
+				b := tempBuf.Bytes()
+				_, err := buf.Write(b)
+				if err == nil {
+					continue
+				}
+				if errors.Is(err, errOverCapacity) {
+					if !buf.Empty() {
+						return iterState{i, j, k, false}, permanentErrors
+					}
+					permanentErrors = append(permanentErrors, consumererror.NewPermanent(
+						fmt.Errorf("dropped metric event: error: event size %d bytes larger than configured max"+
+							" content length %d bytes", len(b), c.config.MaxContentLengthMetrics)))
+					return iterState{i, j, k + 1, false}, permanentErrors
+				}
+				permanentErrors = append(permanentErrors, consumererror.NewPermanent(fmt.Errorf(
+					"error writing the event: %w", err)))
 			}
 		}
 	}
+
+	return iterState{done: true}, permanentErrors
+}
+
+func (c *client) fillMetricsBufferMultiMetrics(events []*splunk.Event, buf buffer, is iterState) (iterState, []error) {
+	var permanentErrors []error
+	jsonStream := jsonStreamPool.Get().(*jsoniter.Stream)
+	defer jsonStreamPool.Put(jsonStream)
 
 	for i := is.record; i < len(events); i++ {
 		event := events[i]
@@ -283,66 +323,6 @@ func (c *client) fillMetricsBuffer(metrics pmetric.Metrics, buf buffer, is iterS
 		} else if err != nil {
 			permanentErrors = append(permanentErrors, consumererror.NewPermanent(fmt.Errorf(
 				"error writing the event: %w", err)))
-
-		}
-	}
-
-	return iterState{done: true}, permanentErrors
-}
-
-func (c *client) fillMetricsBufferMultiMetrics(metrics pmetric.Metrics, buf buffer, is iterState) (iterState, []error) {
-	var permanentErrors []error
-	jsonStream := jsonStreamPool.Get().(*jsoniter.Stream)
-	defer jsonStreamPool.Put(jsonStream)
-
-	var events []*splunk.Event
-
-	for i := 0; i < metrics.ResourceMetrics().Len(); i++ {
-		rm := metrics.ResourceMetrics().At(i)
-		for j := 0; j < rm.ScopeMetrics().Len(); j++ {
-			sm := rm.ScopeMetrics().At(j)
-			for k := 0; k < sm.Metrics().Len(); k++ {
-				metric := sm.Metrics().At(k)
-
-				// Parsing metric record to Splunk event.
-				events = append(events, mapMetricToSplunkEvent(rm.Resource(), metric, c.config, c.logger)...)
-			}
-		}
-	}
-
-	merged, err := mergeEventsToMultiMetricFormat(events)
-	if err != nil {
-		permanentErrors = append(permanentErrors, consumererror.NewPermanent(fmt.Errorf(
-			"error merging events: %w", err)))
-	}
-
-	for i := is.record; i < len(merged); i++ {
-		event := merged[i]
-		// JSON encoding event and writing to buffer.
-		b, jsonErr := marshalEvent(event, c.config.MaxEventSize, jsonStream)
-		if jsonErr != nil {
-			permanentErrors = append(permanentErrors, consumererror.NewPermanent(fmt.Errorf("dropped metric event: %v, error: %w", event, jsonErr)))
-			continue
-		}
-		_, err = buf.Write(b)
-		if errors.Is(err, errOverCapacity) {
-			if !buf.Empty() {
-				return iterState{
-					record: i,
-					done:   false,
-				}, permanentErrors
-			}
-			permanentErrors = append(permanentErrors, consumererror.NewPermanent(
-				fmt.Errorf("dropped metric event: error: event size %d bytes larger than configured max"+
-					" content length %d bytes", len(b), c.config.MaxContentLengthMetrics)))
-			return iterState{
-				record: i + 1,
-				done:   i+1 != len(merged),
-			}, permanentErrors
-		} else if err != nil {
-			permanentErrors = append(permanentErrors, consumererror.NewPermanent(fmt.Errorf(
-				"error writing the event: %w", err)))
-
 		}
 	}
 
@@ -396,6 +376,53 @@ func (c *client) fillTracesBuffer(traces ptrace.Traces, buf buffer, is iterState
 	return iterState{done: true}, permanentErrors
 }
 
+// pushMultiMetricsDataInBatches sends batches of Splunk multi-metric events in JSON format.
+// The batch content length is restricted to MaxContentLengthMetrics.
+// md metrics are parsed to Splunk events.
+func (c *client) pushMultiMetricsDataInBatches(ctx context.Context, md pmetric.Metrics, headers map[string]string) error {
+	buf := c.bufferPool.get()
+	defer c.bufferPool.put(buf)
+	is := iterState{}
+
+	var permanentErrors []error
+	var events []*splunk.Event
+	for i := 0; i < md.ResourceMetrics().Len(); i++ {
+		rm := md.ResourceMetrics().At(i)
+		for j := 0; j < rm.ScopeMetrics().Len(); j++ {
+			sm := rm.ScopeMetrics().At(j)
+			for k := 0; k < sm.Metrics().Len(); k++ {
+				metric := sm.Metrics().At(k)
+
+				// Parsing metric record to Splunk event.
+				events = append(events, mapMetricToSplunkEvent(rm.Resource(), metric, c.config, c.logger)...)
+			}
+		}
+	}
+
+	merged, err := mergeEventsToMultiMetricFormat(events)
+	if err != nil {
+		return consumererror.NewPermanent(fmt.Errorf("error merging events: %w", err))
+	}
+
+	for !is.done {
+		buf.Reset()
+		var latestIterState iterState
+		var batchPermanentErrors []error
+
+		latestIterState, batchPermanentErrors = c.fillMetricsBufferMultiMetrics(merged, buf, is)
+		permanentErrors = append(permanentErrors, batchPermanentErrors...)
+		if !buf.Empty() {
+			if err := c.postEvents(ctx, buf, headers); err != nil {
+				return consumererror.NewMetrics(err, md)
+			}
+		}
+
+		is = latestIterState
+	}
+
+	return multierr.Combine(permanentErrors...)
+}
+
 // pushMetricsDataInBatches sends batches of Splunk events in JSON format.
 // The batch content length is restricted to MaxContentLengthMetrics.
 // md metrics are parsed to Splunk events.
@@ -409,17 +436,14 @@ func (c *client) pushMetricsDataInBatches(ctx context.Context, md pmetric.Metric
 		buf.Reset()
 		var latestIterState iterState
 		var batchPermanentErrors []error
-		if c.config.UseMultiMetricFormat {
-			latestIterState, batchPermanentErrors = c.fillMetricsBufferMultiMetrics(md, buf, is)
-		} else {
-			latestIterState, batchPermanentErrors = c.fillMetricsBuffer(md, buf, is)
-		}
+		latestIterState, batchPermanentErrors = c.fillMetricsBuffer(md, buf, is)
 		permanentErrors = append(permanentErrors, batchPermanentErrors...)
 		if !buf.Empty() {
 			if err := c.postEvents(ctx, buf, headers); err != nil {
-				return consumererror.NewMetrics(err, md)
+				return consumererror.NewMetrics(err, subMetrics(md, is))
 			}
 		}
+
 		is = latestIterState
 	}
 
@@ -493,6 +517,49 @@ func subLogs(src plog.Logs, state iterState) plog.Logs {
 			}
 			for ; k < logs.Len(); k++ {
 				logs.At(k).CopyTo(logsSub.AppendEmpty())
+			}
+		}
+	}
+
+	return dst
+}
+
+// subMetrics returns a subset of metrics starting from the state.
+func subMetrics(src pmetric.Metrics, state iterState) pmetric.Metrics {
+	if state.empty() {
+		return src
+	}
+
+	dst := pmetric.NewMetrics()
+	resources := src.ResourceMetrics()
+	resourcesSub := dst.ResourceMetrics()
+
+	for i := state.resource; i < resources.Len(); i++ {
+		newSub := resourcesSub.AppendEmpty()
+		resources.At(i).Resource().CopyTo(newSub.Resource())
+
+		libraries := resources.At(i).ScopeMetrics()
+		librariesSub := newSub.ScopeMetrics()
+
+		j := 0
+		if i == state.resource {
+			j = state.library
+		}
+		for ; j < libraries.Len(); j++ {
+			lib := libraries.At(j)
+
+			newLibSub := librariesSub.AppendEmpty()
+			lib.Scope().CopyTo(newLibSub.Scope())
+
+			metrics := lib.Metrics()
+			metricsSub := newLibSub.Metrics()
+
+			k := 0
+			if i == state.resource && j == state.library {
+				k = state.record
+			}
+			for ; k < metrics.Len(); k++ {
+				metrics.At(k).CopyTo(metricsSub.AppendEmpty())
 			}
 		}
 	}