… exporter/splunkhecexporter/client_test.go