**Description:**
Apply multi-metric merge at the level of the whole batch rather than within events emitted for one metric. 

**Link to tracking Issue:**
Fixes #23365

**Testing:**
Existing tests pass. Added benchmarks as well for metrics. Added a few tests to make sure we cover the code paths involved.

**Documentation:**
N/A