@atoulme 2023-06-14
before
```
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/splunkhecexporter
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
Benchmark_pushLogData_10_10_1024
Benchmark_pushLogData_10_10_1024-16                              	    5700	    226037 ns/op	   84959 B/op	    1284 allocs/op
Benchmark_pushLogData_10_10_8K
Benchmark_pushLogData_10_10_8K-16                                	    6512	    189749 ns/op	   74050 B/op	    1119 allocs/op
Benchmark_pushLogData_10_10_2M
Benchmark_pushLogData_10_10_2M-16                                	    7270	    171174 ns/op	   73381 B/op	    1108 allocs/op
Benchmark_pushLogData_10_200_2M
Benchmark_pushLogData_10_200_2M-16                               	     321	   3438095 ns/op	 1460949 B/op	   22010 allocs/op
Benchmark_pushLogData_100_200_2M
Benchmark_pushLogData_100_200_2M-16                              	      36	  32864732 ns/op	14914688 B/op	  220024 allocs/op
Benchmark_pushLogData_100_200_5M
Benchmark_pushLogData_100_200_5M-16                              	      36	  35795721 ns/op	14913414 B/op	  220012 allocs/op
Benchmark_pushLogData_compressed_10_10_1024
Benchmark_pushLogData_compressed_10_10_1024-16                   	     967	   1183632 ns/op	 3078518 B/op	    1192 allocs/op
Benchmark_pushLogData_compressed_10_10_8K
Benchmark_pushLogData_compressed_10_10_8K-16                     	    3996	    302914 ns/op	   75552 B/op	    1108 allocs/op
Benchmark_pushLogData_compressed_10_10_2M
Benchmark_pushLogData_compressed_10_10_2M-16                     	    4258	    382858 ns/op	   75217 B/op	    1108 allocs/op
Benchmark_pushLogData_compressed_10_200_2M
Benchmark_pushLogData_compressed_10_200_2M-16                    	     164	   6869773 ns/op	 1463430 B/op	   22010 allocs/op
Benchmark_pushLogData_compressed_100_200_2M
Benchmark_pushLogData_compressed_100_200_2M-16                   	      16	  68470229 ns/op	14632091 B/op	  220014 allocs/op
Benchmark_pushLogData_compressed_100_200_5M
Benchmark_pushLogData_compressed_100_200_5M-16                   	      18	  68477824 ns/op	14624535 B/op	  220014 allocs/op
Benchmark_pushMetricData_10_10_1024
Benchmark_pushMetricData_10_10_1024-16                           	    3907	    304551 ns/op	  188901 B/op	    1908 allocs/op
Benchmark_pushMetricData_10_10_8K
Benchmark_pushMetricData_10_10_8K-16                             	    2546	    505665 ns/op	  907662 B/op	    1909 allocs/op
Benchmark_pushMetricData_10_10_2M
Benchmark_pushMetricData_10_10_2M-16                             	      48	  31726194 ns/op	209807616 B/op	    1920 allocs/op
Benchmark_pushMetricData_10_200_2M
Benchmark_pushMetricData_10_200_2M-16                            	       1	1335663317 ns/op	4197092192 B/op	   38181 allocs/op
Benchmark_pushMetricData_100_200_2M
Benchmark_pushMetricData_100_200_2M-16                           	       1	7007132288 ns/op	41966504088 B/op	  380328 allocs/op
Benchmark_pushMetricData_100_200_5M
Benchmark_pushMetricData_100_200_5M-16                           	       1	13063212208 ns/op	104884244792 B/op	  380593 allocs/op
Benchmark_pushMetricData_compressed_10_10_1024
Benchmark_pushMetricData_compressed_10_10_1024-16                	    1935	    600858 ns/op	  192530 B/op	    1908 allocs/op
Benchmark_pushMetricData_compressed_10_10_8K
Benchmark_pushMetricData_compressed_10_10_8K-16                  	    1406	    782208 ns/op	  916616 B/op	    1909 allocs/op
Benchmark_pushMetricData_compressed_10_10_2M
Benchmark_pushMetricData_compressed_10_10_2M-16                  	      54	  19117786 ns/op	209957431 B/op	    1919 allocs/op
Benchmark_pushMetricData_compressed_10_200_2M
Benchmark_pushMetricData_compressed_10_200_2M-16                 	       3	 430666196 ns/op	4196322442 B/op	   38072 allocs/op
Benchmark_pushMetricData_compressed_100_200_2M
Benchmark_pushMetricData_compressed_100_200_2M-16                	       1	4707735650 ns/op	41961542576 B/op	  380279 allocs/op
Benchmark_pushMetricData_compressed_100_200_5M
Benchmark_pushMetricData_compressed_100_200_5M-16                	       1	10964288560 ns/op	104876135600 B/op	  380485 allocs/op
Benchmark_pushMetricData_10_10_1024_MultiMetric
Benchmark_pushMetricData_10_10_1024_MultiMetric-16               	    1599	    656861 ns/op	  341616 B/op	    4110 allocs/op
Benchmark_pushMetricData_10_10_8K_MultiMetric
Benchmark_pushMetricData_10_10_8K_MultiMetric-16                 	    1394	    835416 ns/op	 1061427 B/op	    4114 allocs/op
Benchmark_pushMetricData_10_10_2M_MultiMetric
Benchmark_pushMetricData_10_10_2M_MultiMetric-16                 	      43	  23478955 ns/op	210132618 B/op	    4423 allocs/op
Benchmark_pushMetricData_10_200_2M_MultiMetric
Benchmark_pushMetricData_10_200_2M_MultiMetric-16                	       3	 425045247 ns/op	4201105288 B/op	   85641 allocs/op
Benchmark_pushMetricData_100_200_2M_MultiMetric
Benchmark_pushMetricData_100_200_2M_MultiMetric-16               	       1	6498838607 ns/op	42005825824 B/op	  839158 allocs/op
Benchmark_pushMetricData_100_200_5M_MultiMetric
Benchmark_pushMetricData_100_200_5M_MultiMetric-16               	       1	15732649006 ns/op	104933503528 B/op	  858013 allocs/op
Benchmark_pushMetricData_compressed_10_10_1024_MultiMetric
Benchmark_pushMetricData_compressed_10_10_1024_MultiMetric-16    	     884	   1161600 ns/op	  342528 B/op	    4110 allocs/op
Benchmark_pushMetricData_compressed_10_10_8K_MultiMetric
Benchmark_pushMetricData_compressed_10_10_8K_MultiMetric-16      	     849	   1413577 ns/op	 1068147 B/op	    4116 allocs/op
Benchmark_pushMetricData_compressed_10_10_2M_MultiMetric
Benchmark_pushMetricData_compressed_10_10_2M_MultiMetric-16      	      43	  24563321 ns/op	210113655 B/op	    4359 allocs/op
Benchmark_pushMetricData_compressed_10_200_2M_MultiMetric
Benchmark_pushMetricData_compressed_10_200_2M_MultiMetric-16     	       2	 527254080 ns/op	4201183424 B/op	   85820 allocs/op
Benchmark_pushMetricData_compressed_100_200_2M_MultiMetric
Benchmark_pushMetricData_compressed_100_200_2M_MultiMetric-16    	       1	6749746937 ns/op	42000147704 B/op	  840409 allocs/op
Benchmark_pushMetricData_compressed_100_200_5M_MultiMetric
Benchmark_pushMetricData_compressed_100_200_5M_MultiMetric-16    	       1	16930777867 ns/op	104923245560 B/op	  856262 allocs/op
BenchmarkConsumeLogsRejected
BenchmarkConsumeLogsRejected-16                                  	     511	   2289710 ns/op	  731440 B/op	   11012 allocs/op
```

after
```
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/splunkhecexporter
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
Benchmark_pushLogData_10_10_1024
Benchmark_pushLogData_10_10_1024-16                              	    5840	    192911 ns/op	   84956 B/op	    1284 allocs/op
Benchmark_pushLogData_10_10_8K
Benchmark_pushLogData_10_10_8K-16                                	    7075	    168736 ns/op	   74063 B/op	    1119 allocs/op
Benchmark_pushLogData_10_10_2M
Benchmark_pushLogData_10_10_2M-16                                	    7609	    162157 ns/op	   73393 B/op	    1108 allocs/op
Benchmark_pushLogData_10_200_2M
Benchmark_pushLogData_10_200_2M-16                               	     368	   3217723 ns/op	 1462744 B/op	   22010 allocs/op
Benchmark_pushLogData_100_200_2M
Benchmark_pushLogData_100_200_2M-16                              	      39	  30912133 ns/op	14726611 B/op	  220023 allocs/op
Benchmark_pushLogData_100_200_5M
Benchmark_pushLogData_100_200_5M-16                              	      40	  29474789 ns/op	14878428 B/op	  220012 allocs/op
Benchmark_pushLogData_compressed_10_10_1024
Benchmark_pushLogData_compressed_10_10_1024-16                   	     990	   1192856 ns/op	 3106502 B/op	    1193 allocs/op
Benchmark_pushLogData_compressed_10_10_8K
Benchmark_pushLogData_compressed_10_10_8K-16                     	    4038	    289190 ns/op	   75527 B/op	    1108 allocs/op
Benchmark_pushLogData_compressed_10_10_2M
Benchmark_pushLogData_compressed_10_10_2M-16                     	    4460	    278880 ns/op	   75681 B/op	    1108 allocs/op
Benchmark_pushLogData_compressed_10_200_2M
Benchmark_pushLogData_compressed_10_200_2M-16                    	     186	   5947477 ns/op	 1467341 B/op	   22010 allocs/op
Benchmark_pushLogData_compressed_100_200_2M
Benchmark_pushLogData_compressed_100_200_2M-16                   	      18	  63198575 ns/op	14624551 B/op	  220014 allocs/op
Benchmark_pushLogData_compressed_100_200_5M
Benchmark_pushLogData_compressed_100_200_5M-16                   	      20	  56661060 ns/op	14618530 B/op	  220013 allocs/op
Benchmark_pushMetricData_10_10_1024
Benchmark_pushMetricData_10_10_1024-16                           	    5068	    231455 ns/op	   88213 B/op	    1816 allocs/op
Benchmark_pushMetricData_10_10_8K
Benchmark_pushMetricData_10_10_8K-16                             	    5269	    236731 ns/op	   88238 B/op	    1816 allocs/op
Benchmark_pushMetricData_10_10_2M
Benchmark_pushMetricData_10_10_2M-16                             	    5161	    236832 ns/op	   88225 B/op	    1816 allocs/op
Benchmark_pushMetricData_10_200_2M
Benchmark_pushMetricData_10_200_2M-16                            	     258	   4892931 ns/op	 1782614 B/op	   36024 allocs/op
Benchmark_pushMetricData_100_200_2M
Benchmark_pushMetricData_100_200_2M-16                           	      18	  68900785 ns/op	30840724 B/op	  620061 allocs/op
Benchmark_pushMetricData_100_200_5M
Benchmark_pushMetricData_100_200_5M-16                           	      16	  63392414 ns/op	30869394 B/op	  620061 allocs/op
Benchmark_pushMetricData_compressed_10_10_1024
Benchmark_pushMetricData_compressed_10_10_1024-16                	    2491	    450391 ns/op	   89157 B/op	    1816 allocs/op
Benchmark_pushMetricData_compressed_10_10_8K
Benchmark_pushMetricData_compressed_10_10_8K-16                  	    2596	    462666 ns/op	   88805 B/op	    1816 allocs/op
Benchmark_pushMetricData_compressed_10_10_2M
Benchmark_pushMetricData_compressed_10_10_2M-16                  	    2558	    482976 ns/op	   88493 B/op	    1816 allocs/op
Benchmark_pushMetricData_compressed_10_200_2M
Benchmark_pushMetricData_compressed_10_200_2M-16                 	     100	  10061816 ns/op	 1783159 B/op	   36024 allocs/op
Benchmark_pushMetricData_compressed_100_200_2M
Benchmark_pushMetricData_compressed_100_200_2M-16                	      12	  92486088 ns/op	17896480 B/op	  360035 allocs/op
Benchmark_pushMetricData_compressed_100_200_5M
Benchmark_pushMetricData_compressed_100_200_5M-16                	      12	 104681386 ns/op	17896126 B/op	  360035 allocs/op
Benchmark_pushMetricData_10_10_1024_MultiMetric
Benchmark_pushMetricData_10_10_1024_MultiMetric-16               	    1981	    565854 ns/op	  245790 B/op	    3933 allocs/op
Benchmark_pushMetricData_10_10_8K_MultiMetric
Benchmark_pushMetricData_10_10_8K_MultiMetric-16                 	    1917	    574839 ns/op	  245677 B/op	    3933 allocs/op
Benchmark_pushMetricData_10_10_2M_MultiMetric
Benchmark_pushMetricData_10_10_2M_MultiMetric-16                 	    1929	    567547 ns/op	  245759 B/op	    3933 allocs/op
Benchmark_pushMetricData_10_200_2M_MultiMetric
Benchmark_pushMetricData_10_200_2M_MultiMetric-16                	     104	  11053750 ns/op	 5001007 B/op	   78103 allocs/op
Benchmark_pushMetricData_100_200_2M_MultiMetric
Benchmark_pushMetricData_100_200_2M_MultiMetric-16               	       6	 197923046 ns/op	94994392 B/op	 1461163 allocs/op
Benchmark_pushMetricData_100_200_5M_MultiMetric
Benchmark_pushMetricData_100_200_5M_MultiMetric-16               	       4	 284557776 ns/op	95342354 B/op	 1461149 allocs/op
Benchmark_pushMetricData_compressed_10_10_1024_MultiMetric
Benchmark_pushMetricData_compressed_10_10_1024_MultiMetric-16    	    1095	   1101175 ns/op	  247239 B/op	    3934 allocs/op
Benchmark_pushMetricData_compressed_10_10_8K_MultiMetric
Benchmark_pushMetricData_compressed_10_10_8K_MultiMetric-16      	    1346	    921537 ns/op	  246314 B/op	    3934 allocs/op
Benchmark_pushMetricData_compressed_10_10_2M_MultiMetric
Benchmark_pushMetricData_compressed_10_10_2M_MultiMetric-16      	    1348	   1022720 ns/op	  246324 B/op	    3934 allocs/op
Benchmark_pushMetricData_compressed_10_200_2M_MultiMetric
Benchmark_pushMetricData_compressed_10_200_2M_MultiMetric-16     	      62	  18006845 ns/op	 5004480 B/op	   78101 allocs/op
Benchmark_pushMetricData_compressed_100_200_2M_MultiMetric
Benchmark_pushMetricData_compressed_100_200_2M_MultiMetric-16    	       6	 178164323 ns/op	49852669 B/op	  780587 allocs/op
Benchmark_pushMetricData_compressed_100_200_5M_MultiMetric
Benchmark_pushMetricData_compressed_100_200_5M_MultiMetric-16    	       7	 156752667 ns/op	49821093 B/op	  780586 allocs/op
BenchmarkConsumeLogsRejected
BenchmarkConsumeLogsRejected-16                                  	     666	   1797371 ns/op	  732812 B/op	   11012 allocs/op
PASS
```